Moved Helper to own file
